+++ src/chem/aeronet_aod2ioda.py
@@ -142,11 +142,11 @@ if __name__ == '__main__':
         varAttrs[key, iconv.OqcName()]['units'] = ''
 
     for key, value in obsvars.items():
-        outdata[varDict[key]['valKey']] = np.array(f3[value].fillna(np.float32(-999.)))
+        outdata[varDict[key]['valKey']] = np.array(np.float32(f3[value].fillna(np.float32(-999.))))
         outdata[varDict[key]['qcKey']] = np.where(outdata[varDict[key]['valKey']] == np.float32(-999.),
                                                   1, 0)
         outdata[varDict[key]['errKey']] = np.where(outdata[varDict[key]['valKey']] == np.float32(-999.),
-                                                   np.float32(-999.), 0.02)
+                                                   np.float32(-999.), np.float32(0.02))
 
     # Add metadata variables
     outdata[('latitude', 'MetaData')] = np.array(np.float32(f3['latitude']))
